+++ reporting-service/static/expense.js
@@ -51,7 +51,7 @@ function editExpense(id, category, amount, date) {
 }
 
 function deleteExpense(id) {
-    if (confirm("Are you sure you want to delete this expense?")) {
+    if (confirm("Are you sure you want to delete this expense?\n\nWarning: This action is irreversible. Once deleted, the expense data will be permanently lost and cannot be recovered.")) {
         fetch(`/expenses/${id}`, {
             method: 'DELETE',
         })
